requirements.txt
streamlit
pandas
numpy
matplotlib
lightgbm
scikit-learn
joblib
plotly
altair
requests
pillow

seaborn
nltk
statsmodels
optuna
shap
category_encoders
xgboost
catboost
transformers
torch
pytest
ruff
pre-commit
